build: enable cross-platform build & run on Windows / macOS / Linux
- CMakeLists.txt: opt into FetchContent for glfw/assimp on non-Windows
  with system-package fallback; vendored MSVC libs remain default on
  Windows. Trim assimp to OBJ-only and disable exporters so the source
  build stays small and avoids unrelated compile failures.
- common/tools.{h,cpp}: enable GLFW_OPENGL_FORWARD_COMPAT (required for
  the 3.3 Core context on macOS) and add a portable
  anchorWorkingDirectoryToExecutable() helper so "../res/..." resolves
  no matter where the user launches from.
- main.cpp: chdir to the executable's directory before any resource load.
- common/World.h: fix #include <Player.h> -> "player.h" for case-sensitive
  filesystems.
- common/ghost.h, common/player.h: replace non-standard std::sinf with
  std::sin so the code compiles under GCC/Clang.
- .github/workflows/build.yml: matrix CI for Ubuntu / macOS / Windows.
- README.md: rewritten build section with per-platform instructions and
  the new CMake options.

+++ common/player.h
@@ -3,6 +3,7 @@
 #include <glm/glm.hpp>
 #include <glm/gtc/matrix_transform.hpp>
 #include <vector>
+#include <cmath>
 #include <model/model.h>
 
 //Add jump function
@@ -233,7 +234,7 @@ public:
 		modelMat4 = glm::translate(modelMat4, offset);
 		if (glfwGetMouseButton(window, GLFW_MOUSE_BUTTON_1) == GLFW_PRESS)
 		{
-			modelMat4 = glm::translate(modelMat4, glm::vec3(1, 0.4, 0.4) * (std::sinf(6 * glfwGetTime())));
+			modelMat4 = glm::translate(modelMat4, glm::vec3(1, 0.4, 0.4) * static_cast<float>(std::sin(6.0 * glfwGetTime())));
 		}
